Move repository_util function into legacy code in only place used.
Eliminates a galaxy -> tool shed import and centralizes legacy code.
