+++ lib/tool_shed/util/repository_util.py
@@ -363,63 +363,6 @@ def get_repositories_by_category(
     return repositories
 
 
-def get_tool_shed_repository_status_label(
-    app, tool_shed_repository=None, name=None, owner=None, changeset_revision=None, repository_clone_url=None
-):
-    """Return a color-coded label for the status of the received tool-shed_repository installed into Galaxy."""
-    if tool_shed_repository is None:
-        if name is not None and owner is not None and repository_clone_url is not None:
-            tool_shed = get_tool_shed_from_clone_url(repository_clone_url)
-            tool_shed_repository = get_installed_repository(
-                app, tool_shed=tool_shed, name=name, owner=owner, installed_changeset_revision=changeset_revision
-            )
-    if tool_shed_repository:
-        status_label = tool_shed_repository.status
-        if tool_shed_repository.status in [
-            app.install_model.ToolShedRepository.installation_status.CLONING,
-            app.install_model.ToolShedRepository.installation_status.SETTING_TOOL_VERSIONS,
-            app.install_model.ToolShedRepository.installation_status.INSTALLING_REPOSITORY_DEPENDENCIES,
-            app.install_model.ToolShedRepository.installation_status.INSTALLING_TOOL_DEPENDENCIES,
-            app.install_model.ToolShedRepository.installation_status.LOADING_PROPRIETARY_DATATYPES,
-        ]:
-            bgcolor = app.install_model.ToolShedRepository.states.INSTALLING
-        elif tool_shed_repository.status in [
-            app.install_model.ToolShedRepository.installation_status.NEW,
-            app.install_model.ToolShedRepository.installation_status.UNINSTALLED,
-        ]:
-            bgcolor = app.install_model.ToolShedRepository.states.UNINSTALLED
-        elif tool_shed_repository.status in [app.install_model.ToolShedRepository.installation_status.ERROR]:
-            bgcolor = app.install_model.ToolShedRepository.states.ERROR
-        elif tool_shed_repository.status in [app.install_model.ToolShedRepository.installation_status.DEACTIVATED]:
-            bgcolor = app.install_model.ToolShedRepository.states.WARNING
-        elif tool_shed_repository.status in [app.install_model.ToolShedRepository.installation_status.INSTALLED]:
-            if tool_shed_repository.repository_dependencies_being_installed:
-                bgcolor = app.install_model.ToolShedRepository.states.WARNING
-                status_label = "{}, {}".format(
-                    status_label,
-                    app.install_model.ToolShedRepository.installation_status.INSTALLING_REPOSITORY_DEPENDENCIES,
-                )
-            elif tool_shed_repository.missing_repository_dependencies:
-                bgcolor = app.install_model.ToolShedRepository.states.WARNING
-                status_label = f"{status_label}, missing repository dependencies"
-            elif tool_shed_repository.tool_dependencies_being_installed:
-                bgcolor = app.install_model.ToolShedRepository.states.WARNING
-                status_label = "{}, {}".format(
-                    status_label, app.install_model.ToolShedRepository.installation_status.INSTALLING_TOOL_DEPENDENCIES
-                )
-            elif tool_shed_repository.missing_tool_dependencies:
-                bgcolor = app.install_model.ToolShedRepository.states.WARNING
-                status_label = f"{status_label}, missing tool dependencies"
-            else:
-                bgcolor = app.install_model.ToolShedRepository.states.OK
-        else:
-            bgcolor = app.install_model.ToolShedRepository.states.ERROR
-    else:
-        bgcolor = app.install_model.ToolShedRepository.states.WARNING
-        status_label = "unknown status"
-    return f'<div class="count-box state-color-{bgcolor}">{status_label}</div>'
-
-
 def handle_role_associations(app, role, repository, **kwd):
     sa_session = app.model.session
     message = escape(kwd.get("message", ""))
